ported more analysis_utils functions
