+++ calcification/analysis/analysis.py
@@ -12,8 +12,6 @@ import rpy2.robjects.packages as rpackages
 # stats
 import statsmodels.api as sm
 from rpy2.robjects import pandas2ri
-from scipy.stats import median_abs_deviation
-from scipy.stats import norm as scipy_norm
 from tqdm.auto import tqdm
 
 from calcification.analysis import analysis_utils
@@ -1195,533 +1193,6 @@ def generate_location_specific_predictions(
         metadata_rows[i]["predicted_effect_size"] = val
 
     return metadata_rows
-
-
-def filter_robust_zscore(series: pd.Series, threshold: float = 20) -> pd.Series:
-    """
-    Filter out outliers based on robust z-scores.
-
-    Args:
-        series (pd.Series): The series to filter.
-        threshold (float): The z-score threshold for filtering.
-
-    Returns:
-        pd.Series: A boolean series indicating which values are not outliers.
-    """
-    median = np.median(series)
-    mad = median_abs_deviation(
-        series, scale="normal"
-    )  # scale for approx equivalence to std dev
-    robust_z = np.abs((series - median) / mad)
-    return robust_z < threshold
-
-
-def p_score(prediction: float, se: float, null_value: float = 0) -> float:
-    """
-    Calculate the p-value for a given prediction and standard error.
-    """
-    z = (prediction - null_value) / se
-    p = 2 * (1 - scipy_norm.cdf(abs(z)))  # two-tailed p-value
-    return p
-
-
-### assign certainty levels
-def assign_certainty(p_score: float) -> int:
-    """
-    Assign certainty levels based on p-value.
-    """
-    if p_score < 0.01:
-        return 4  # very high certainty
-    elif p_score < 0.05:
-        return 3  # high certainty
-    elif p_score < 0.1:
-        return 2  # medium certainty
-    else:
-        return 1  # low certainty
-
-
-### DEPRECATED
-# def compute_heds
-# def calculate_effect_size(df1_sample, df2_sample, var, group1, group2):
-#     other_var = 'temp' if var == 'phtot' else 'phtot'
-#     if np.isclose(df1_sample[other_var], df2_sample[other_var], atol=0.1):
-#         if (var == 'phtot' and df1_sample[var] < df2_sample[var]) or (var == 'temp' and df1_sample[var] > df2_sample[var]):
-#             group1, group2 = group2, group1
-#             df1_sample, df2_sample = df2_sample, df1_sample
-#         delta_var = abs(df2_sample[var] - df1_sample[var])
-#         mu1, std1, n1 = df1_sample['calcification'], df1_sample['calcification_sd'], df1_sample['n']
-#         mu2, std2, n2 = df2_sample['calcification'], df2_sample['calcification_sd'], df2_sample['n']
-#         g = calc_hedges_g(mu2, mu1, std2, std1, n2, n1)
-#         return {
-#             'doi': df.doi.iloc[0],
-#             'species_types': species,
-#             'group1': group1,
-#             'group2': group2,
-#             'variable': var,
-#             'delta_var': delta_var,
-#             'hedges_g': g[0],
-#             'hg_ci_l': g[1][0],
-#             'hg_ci_u': g[1][1],
-#             'control_val': df1_sample[var],
-#             'treatment_val': df2_sample[var],
-#         }
-#     return None
-
-# results = []
-# for species, group_df in df.groupby('species_types'):
-#     treatment_groups = group_df['treatment_group'].unique()
-#     for group1, group2 in combinations(treatment_groups, 2):
-#         df1 = group_df[group_df['treatment_group'] == group1]
-#         df2 = group_df[group_df['treatment_group'] == group2]
-#         if df1.n.all() == 1 and df2.n.all() == 1:
-#             df1['calcification_sd'] = np.std(df1['calcification'])
-#             df2['calcification_sd'] = np.std(df2['calcification'])
-#             n1, n2 = len(df1), len(df2)
-#             df1 = utils.aggregate_df(df1)
-#             df2 = utils.aggregate_df(df2)
-#             df1['n'] = n1
-#             df2['n'] = n2
-#         if len(df1) != len(df2):
-#             print(f"Skipping comparison between {group1} and {group2} treatments: Different sample sizes")
-#             continue
-#         for sample in range(len(df1)):
-#             df1_sample = df1.iloc[sample] if isinstance(df1, pd.DataFrame) else df1
-#             df2_sample = df2.iloc[sample] if isinstance(df2, pd.DataFrame) else df2
-#             for var in vars_to_compare:
-#                 result = calculate_effect_size(df1_sample, df2_sample, var, group1, group2)
-#                 if result:
-#                     results.append(result)
-# return results
-
-
-# def compute_hedges_g(df, vars_to_compare=['temp', 'phtot', 'irr']):
-#     """
-#     Compute Hedges' g effect size for each treatment compared to the control within each species.
-
-#     Args:
-#         df (pd.DataFrame): DataFrame containing treatment groups.
-#         vars_to_compare (list): List of numeric variables to compute Hedges' g for.
-
-#     Returns:
-#         pd.DataFrame: DataFrame containing effect sizes for each treatment group.
-#     """
-#     cols = vars_to_compare + ['species_types', 'treatment_group', 'calcification', 'n']
-#     results = []
-#     # remove any vars to compare that are not in the df (accounts for singular response studies)
-#     df['multi_var'] = 1
-#     # remove any of vars_to_compare columns which are all nan
-#     specific_vars_to_compare = [col for col in vars_to_compare if df[col].notna().all()]
-
-#     # if vars_to_compare updated, print message
-#     if len(specific_vars_to_compare) != len(vars_to_compare):
-#         print(f"Vars to compare updated to {specific_vars_to_compare} for {df.doi.iloc[0]}")
-#         # update value of 'multi-var' column to 1
-#         df['multi_var'] = 0
-
-#     for species, group_df in df.groupby('species_types'):
-#         treatment_groups = group_df['treatment_group'].unique()
-#         for tg1, tg2 in combinations(treatment_groups, 2):
-#             df1 = group_df[group_df['treatment_group'] == tg1]
-#             df2 = group_df[group_df['treatment_group'] == tg2]
-
-#             if len(df1) != len(df2):
-#                 print(f"Skipping comparison between {tg1} and {tg2} treatments: Different sample sizes")
-#                 continue
-
-#             if (df1['n'] == 1).all() and(df2['n'] == 1).all():
-#                 df1['calcification_sd'] = np.std(df1['calcification'])
-#                 df2['calcification_sd'] = np.std(df2['calcification'])
-#                 n1, n2 = len(df1), len(df2)
-#                 df1 = utils.aggregate_df(df1)
-#                 df2 = utils.aggregate_df(df2)
-#                 df1['n'] = n1
-#                 df2['n'] = n2
-
-#             for sample in range(len(df1)):
-#                 df1_sample = df1.iloc[sample] if isinstance(df1, pd.DataFrame) else df1
-#                 df2_sample = df2.iloc[sample] if isinstance(df2, pd.DataFrame) else df2
-#                 for var in specific_vars_to_compare:
-#                     other_var = 'temp' if var == 'phtot' else 'phtot'
-#                     if np.isclose(df1_sample[other_var], df2_sample[other_var], atol=0.1):  # if agree within 0.1, likely control variable
-#                         # ensure that df1_sample is control group (smaller value for t_in, larger value for phtot). # TODO: allow switching for deltavar relative to climatology
-#                         if (var == 'phtot' and df1_sample[var] < df2_sample[var]) or (var == 'temp' and df1_sample[var] > df2_sample[var]):
-#                             tg1, tg2 = tg2, tg1
-#                             df1_sample, df2_sample = df2_sample, df1_sample
-#                         delta_var = abs(df2_sample[var] - df1_sample[var])
-#                         print('df1_sample', df1_sample[cols])
-#                         print('df2_sample', df2_sample[cols])
-#                         mu1, std1, n1 = df1_sample['calcification'], df1_sample['calcification_sd'], df1_sample['n']    # control
-#                         mu2, std2, n2 = df2_sample['calcification'], df2_sample['calcification_sd'], df2_sample['n']    # treatment
-#                         g = calc_hedges_g(mu2, mu1, std2, std1, n2, n1)
-#                         results.append({
-#                             'doi': df.doi.iloc[0],
-#                             'location': df.location.iloc[0],
-#                             'species_types': species,
-#                             'group1': tg1,
-#                             'group2': tg2,
-#                             'variable': var,
-#                             'delta_var': delta_var,
-#                             'hedges_g': g[0],
-#                             'hg_ci_l': g[1][0],
-#                             'hg_ci_u': g[1][1],
-#                             'control_val': df1_sample[var],
-#                             'treatment_val': df2_sample[var],
-#                             'multi_var': df['multi_var'].iloc[0]
-#                         })
-#     return results
-
-
-# def cluster_treatments(df, vars_to_cluster):
-#     """Cluster treatments based on independent variables and species types."""
-
-#     import warnings
-#     warnings.filterwarnings("ignore", category=UserWarning, module='sklearn.cluster._kmeans')  # Suppress convergence warnings
-
-#     df = df.copy()  # Avoid modifying the original dataframe
-#     # Ensure 'treatment_group' column exists
-#     df['treatment_group'] = np.nan
-
-#     # Cluster separately for each species type
-#     for species, group_df in df.groupby('species_types'):
-#         treatment_data = group_df[vars_to_cluster].dropna(axis=1)  # Only keep non-missing variables
-
-#         if treatment_data.shape[0] < 2:  # Skip clustering if there's only one sample
-#             print(f"Skipping {species}: Not enough samples for clustering: {len(group_df)}")
-#             continue
-
-#         try:
-#             optimal_k, _ = optimal_kmeans(treatment_data)  # Determine optimal clusters
-#             kmeans = KMeans(n_clusters=optimal_k, random_state=42, n_init=10)
-#             df.loc[group_df.index, 'treatment_group'] = kmeans.fit_predict(treatment_data)
-#         except ValueError:
-#             print(f"Error: Could not cluster {species}, df of length {len(group_df)}")
-#             df.loc[group_df.index, 'treatment_group'] = np.nan
-
-#     return df
-
-
-# def optimal_kmeans(data, max_clusters=8):
-#     best_k = 2  # Minimum sensible number of clusters
-#     best_score = -1
-#     scores = []
-
-#     for k in range(2, min(len(data), max_clusters + 1)):  # Avoid excessive clustering
-#         kmeans = KMeans(n_clusters=k, random_state=42, n_init=max_clusters)
-#         labels = kmeans.fit_predict(data)
-#         score = silhouette_score(data, labels)
-#         scores.append((k, score))
-
-#         if score > best_score:
-#             best_score = score
-#             best_k = k
-
-#     return best_k, scores
-
-
-# def group_irradiance(df, irr_col='irr', atol=30):
-#     """
-#     Assigns an 'irr_group' to values that are within a relative tolerance.
-
-#     Args:
-#         df (pd.DataFrame): Input dataframe with an 'irr' column.
-#         irr_col (str): Column name for irradiance values.
-#         rtol (float): Relative tolerance (e.g., 0.10 for 10%).
-
-#     Returns:
-#         pd.DataFrame: Dataframe with new 'irr_group' column.
-#     """
-#     df = df.copy().sort_values(by=irr_col)  # copy to prevent modification of original, sort for efficiency
-#     irr_groups = np.zeros(len(df))  # initialize array for group assignments
-#     group_id = 0
-#     prev_irr = None
-
-#     for i, irr in enumerate(df[irr_col]):
-#         if np.isnan(irr):   # handling grouping where irr is NaN
-#             irr_groups[i] = -1  # Use a distinct value for NaN irradiance
-#         elif prev_irr is None:
-#             # First non-NaN value starts group 0
-#             group_id = 0
-#             irr_groups[i] = group_id
-#         elif np.abs(irr - prev_irr) > atol:
-#             # Outside tolerance, create new group
-#             group_id += 1
-#             irr_groups[i] = group_id
-#         else:
-#             # Within tolerance, use current group
-#             irr_groups[i] = group_id
-
-#         # Update prev_irr only for non-NaN values
-#         if not np.isnan(irr):
-#             prev_irr = irr
-
-#     # Add irr_group column to dataframe
-#     df['irr_group'] = irr_groups
-#     return df
-
-
-### Previously part of process_group
-# loi = ['phtot', 'temp', 'calcification', 'calcification_sd', 'n']
-# control_df = species_df[species_df['treatment_group'] == 'cTcP']    # extract control data
-# if control_df.empty:
-#     print(f"Control group not found for {species_df['doi'].iloc[0]}, for \n{species_df[['species_types']+loi].iloc[0]}")
-#     return None
-# if len(control_df) > 1: # if multiple possible controls, aggregate
-#     control_row = aggregate_by_treatment_group(control_df)
-# else:
-#     control_row = control_df.iloc[0]
-
-# # append control_df to result_dfs
-# if not pd.isna(control_row).all():
-#     control_row_df = pd.DataFrame(control_row).T
-
-#     # Explicitly check for all-NA columns and remove them
-#     if not control_row_df.dropna(how="all", axis=1).empty:
-#         result_dfs.append(control_row_df)
-# # process each treatment group for univariate treatments
-# for treatment_group, treatment_df in species_df.groupby('treatment_group'):
-#     if treatment_group in ['uncertain', 'cTcP']:
-#         # doesn't make sense to process
-#         continue
-
-#     # in the case that all n == 1 (individual datapoints), aggregate by treatment_group
-#     if np.all(treatment_df.n == 1):
-#         treatment_df = pd.DataFrame(aggregate_by_treatment_group(treatment_df)).T # hedges_g_for_row requires df input
-
-
-#     if treatment_group == 'tTtP':
-#         ### option 1: calculate each relative to single control (simple) – the max pH and min T for species
-#         # for each row in treatment_Group, calculate hedges_g_for_row and append to result_dfs
-#         if np.any(treatment_df.n == 1):
-#             print('will need to cluster by treatment level')
-#         out = pd.DataFrame(treatment_df.apply(
-#             lambda row: hedges_g_for_row(row, control_row),
-#             axis=1
-#         ))
-#         result_dfs.append(out) if not out.empty else None
-#         ### option 2: for each treatment, calculate each relative to all other treatments (complex)
-
-
-#         # e.g. for pHs at higher T, determine highest pH and calculate effects wrt that
-#         # and for Ts, determine pH closest at a different T and calculate effect wrt that
-#         continue
-
-#     # apply hedges_g calculation to each row
-#     processed_df = treatment_df.apply(
-#         lambda row: hedges_g_for_row(row, control_row),
-#         axis=1
-#     )
-#     result_dfs.append(processed_df) if not processed_df.empty else None
-
-# if result_dfs:
-#     # Drop all-NA columns from each DataFrame before concatenation
-#     filtered_dfs = [df.dropna(how="all", axis=1) for df in result_dfs]
-
-#     # Ensure at least one DataFrame is non-empty before concatenation
-#     filtered_dfs = [df for df in filtered_dfs if not df.empty]
-
-#     return pd.concat(filtered_dfs, axis=0) if filtered_dfs else None
-# else:
-#     return None
-
-
-# def process_group(species_df: pd.DataFrame, effect_type: str = 'hedges_g') -> pd.DataFrame:
-#     """
-#     Process a group of species data to calculate Hedges' g.
-
-#     Args:
-#         species_df: DataFrame containing data for a specific species
-
-#     Returns:
-#         pandas.DataFrame: DataFrame with Hedges' g calculations
-#     """
-#     # TODO: implement simple method (single control): see deprecated code
-#     result_dfs = []
-#     result_dfs.extend(process_group_multivar(species_df, effect_type=effect_type))
-
-#     # suppress futurewarning
-#     import warnings
-#     warnings.simplefilter(action='ignore', category=FutureWarning)
-
-#     if result_dfs:  # TODO: fix future warning without dropping all-nan columns
-#         # Identify all column names across all DataFrames
-#         all_columns = set().union(*(df.columns for df in result_dfs))
-
-#         # Ensure all DataFrames contain the same columns with consistent dtypes
-#         for i, df in enumerate(result_dfs):
-#             missing_cols = all_columns - set(df.columns)
-#             for col in missing_cols:
-#                 df[col] = pd.NA  # Retain columns but explicitly mark as NA
-#             result_dfs[i] = df.astype({col: "object" for col in missing_cols})  # Set dtype to avoid dtype inference issues
-
-#     # Concatenate DataFrames without dropping NaN-only columns
-#         result_df = pd.concat(result_dfs, axis=0, ignore_index=True)
-#         # # drop all-NA columns from each DataFrame before concatenation (avoids future warning)
-#         # filtered_dfs = [df.dropna(how="all", axis=1) for df in result_dfs]
-#         # # ensure at least one DataFrame is non-empty before concatenation
-#         # filtered_dfs = [df for df in filtered_dfs if not df.empty]
-#         return result_df
-#         # return pd.concat(filtered_dfs, axis=0) if filtered_dfs else None
-#     else:
-#         return None
-
-
-# def calc_cohens_d(mu1: float, mu2: float, sd_pooled: float) -> float:
-#     """Calculate Hedges G metric: https://www.itl.nist.gov/div898/software/dataplot/refman2/auxillar/hedgeg.htm
-
-#     Args:
-#         mu1 (float): mean of group 1
-#         mu2 (float): mean of group 2
-#     2 (float): mean of group 2
-#         sd_pooled (float): pooled standard deviation of both groups
-
-#     Returns:
-#         float: Hedges G metric
-#     """
-#     return (mu1 - mu2) / sd_pooled
-
-
-# def calc_sd_pooled(n1: int, n2: int, sd1: float, sd2: float) -> float:
-#     """Calculate pooled standard deviation: https://www.itl.nist.gov/div898/software/dataplot/refman2/auxillar/hedgeg.htm
-
-#     Args:
-#         n1 (int): number of samples in group 1
-#         n2 (int): number of samples in group 2
-#         sd1 (float): standard deviation of group 1
-#         sd2 (float): standard deviation of group 2
-
-#     Returns:
-#         float: pooled standard deviation
-#     """
-#     return np.sqrt(((n1 - 1) * sd1 ** 2 + (n2 - 1) * sd2 ** 2) / (n1 + n2 - 2))
-
-
-# def calc_sd_pooled(n1: int, n2: int, sd1: float, sd2: float) -> float:
-#     """Calculate pooled standard deviation: https://www.itl.nist.gov/div898/software/dataplot/refman2/auxillar/hedgeg.htm
-#     NB the simpler version (as used by Ben). Doesn't weight by sample size (although this slightly captured by sd)
-#     Args:
-#         n1 (int): number of samples in group 1
-#         n2 (int): number of samples in group 2
-#         sd1 (float): standard deviation of group 1
-#         sd2 (float): standard deviation of group 2
-
-#     Returns:
-#         float: pooled standard deviation
-#     """
-#     return np.sqrt((sd1 ** 2 + sd2 ** 2) / 2)
-
-
-# def process_group_multivar(df: pd.DataFrame) -> pd.DataFrame:
-#     """
-#     Process a group of species data to calculate effect size.
-
-#     Args:
-#         df: DataFrame containing data for a specific species
-
-#     Returns:
-#         pandas.DataFrame: DataFrame with effect size calculations
-#     """
-
-#     results_df = []
-#     # for each treatment value in pH, calculate effect size varying temperature
-#     grouped_by_ph = df.groupby('treatment_level_ph')
-#     grouped_by_t = df.groupby('treatment_level_t')
-
-#     def process_group(group, control_level_col):
-#         control_level = min(group[control_level_col])
-#         control_df = group[group[control_level_col] == control_level]
-#         treatment_df = group[group[control_level_col] > control_level]
-
-#         if treatment_df.empty:  # skip if there's no treatment data
-#             return
-
-#         # covered (better) before being passed
-#         # # aggregate control data if n > 1 (more precise than just taking first row)
-#         # control_data = aggregate_by_treatment_group(control_df) if len(control_df) > 1 else control_df.iloc[0]
-#         # # Aggregate treatment data if all n=1
-#         # if np.all(treatment_df.n == 1):
-#         #     treatment_df = pd.DataFrame(aggregate_by_treatment_group(treatment_df)).T
-
-
-#         # update treatment label
-#         if control_level_col == "treatment_level_t":
-#             treatment_df.loc[:, 'treatment'] = 'temp'
-#         elif control_level_col == "treatment_level_ph":
-#             treatment_df.loc[:, 'treatment'] = 'phtot'
-
-#         # calculate effect size varying treatment condition
-#         # First make sure we convert control_data DataFrame to Series if needed
-#         control_series = control_df.iloc[0] if isinstance(control_df, pd.DataFrame) else control_df
-#         # Calculate effect size for each row in treatment_df
-#         effect_size = treatment_df.apply(lambda row: calc_treatment_effect_for_row(row, control_series), axis=1)
-#         return effect_size
-
-#     results_df.append(grouped_by_ph.apply(process_group, control_level_col='treatment_level_t', include_groups=False))
-#     results_df.append(grouped_by_t.apply(process_group, control_level_col='treatment_level_ph', include_groups=False))
-
-#     return pd.concat(results_df).reset_index(drop=True)
-
-
-# # Import required R packages
-# from rpy2.robjects.packages import importr
-# from rpy2.robjects import pandas2ri
-# import rpy2.robjects as ro
-
-# # Activate pandas2ri for DataFrame conversion
-# pandas2ri.activate()
-# r = ro.r
-
-# # Import metafor package
-# metafor = importr('metafor')
-
-# # temp_df['esid'] = temp_df.groupby('doi').cumcount() + 1
-
-# # Convert your DataFrame to an R DataFrame
-# r_temp_df = pandas2ri.py2rpy(temp_df)
-# ro.r.assign('dat', r_temp_df)
-# ro.r.assign('res', basic_temp_model)
-
-
-# # Now you can use the updated 'dat' and 'V' in your forest plot or further analysis
-# ro.r('''
-# create_forest_plot <- function(res, dat) {
-#     par(tck=-.01, mgp=c(1,0.01,0), mar=c(2,4,2,2))
-
-#     # Ensure 'original_doi' is numeric to avoid non-numeric argument errors
-#     dat$original_doi <- as.numeric(as.factor(dat$original_doi))
-#     # sort by original_doi
-#     dat <- dat[order(dat$original_doi), ]
-
-#     dd <- c(0, diff(dat$original_doi))
-#     print(dd)
-#     rows <- (1:res$k)*5 + cumsum(dd)  # Multiply by 2 to increase vertical spacing
-#     forest(res,
-#     # rows=rows,
-#     # ylim=c(-500, max(rows)+500), xlim=c(-5,7),
-#                  cex=0.4, efac=c(0,1), mlab="Pooled Estimate")
-
-#     # abline(h = rows[c(1,diff(rows)) == 2] - 1, lty="dotted")
-# }
-# ''')
-# from rpy2.robjects.packages import importr
-
-# grdevices = importr('grDevices')
-
-# # To execute the forest plot function
-# ro.r('create_forest_plot(res, dat)')
-# # save the plot
-
-# # If you want to save the plot
-# r('pdf("forest_plot.pdf", width=12, height=10)')
-# r('create_forest_plot(res, dat)')
-# r('dev.off()')
-
-# # grdevices.png(str(config.fig_dir / 'forest_plot.png'), width=800, height=600)
-# # grdevices.dev_off()
-# # # view plot in python
-# # from matplotlib import pyplot as plt
-# # import matplotlib.image as mpimg
-# # img = mpimg.imread('forest_plot.png')
-# # plt.imshow(img)
 
 
 class MetaForModel:
